requirements.txt
Django==4.2.4
djangorestframework==3.14.0
drf-yasg==1.21.7
prometheus-client==0.17.1
psycopg2-binary==2.9.7
